requirements.txt
flask
flask_sqlalchemy
psycopg2-binary #si estan en mac o linux psycopg2-binary en windows psycopg2
requests
gunicorn
python-dotenv